requirements.txt
flet~=0.22.0
networkx~=3.2.1
mysql-connector-python~=8.3.0
colorama==0.4.6